% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{applicable_solvers}
\alias{applicable_solvers}
\title{Discover Applicable Solvers}
\usage{
applicable_solvers(.opt)
}
\arguments{
\item{.opt}{An object of the `OP` class.}
}
\value{
A \code{list} of the `OP` class.
}
\description{
Discover Applicable Solvers
}
\examples{
#
}
