% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-snoop.R
\name{autoplot.snoop_statistics}
\alias{autoplot.snoop_statistics}
\title{Autoplot Method}
\usage{
\method{autoplot}{snoop_statistics}(.data)
}
\arguments{
\item{.data}{An object of the \code{snoop_statistics} class.}
}
\value{
A \code{ggplo2} object.
}
\description{
A quick visual inspection of objects from the \code{snoop_statistics} class.
}
\examples{
library(ggplot2)

stocks <- tibble::tibble(
  time = as.Date('2009-01-01') + 0:99,
  X    = rnorm(100, 0, 1),
  Y    = rnorm(100, 0, 2),
  Z    = rnorm(100, 0, 4)
)

roll <- construct_rolling_infrastructure(stocks, .initial = 50)
rebal <- construct_rebalance_infrastructure(roll)

# Risk-Parity Strategy
compute_cov <- function(.data) stats::cov(as.matrix(.data))

rebalance_portfolio(rebal, compute_cov, .strategy = "risk_parity") |>
  extract_statistics() |>
  autoplot()
}
