% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_bounds}
\alias{constraint_bounds}
\alias{constraint_bounds.default}
\alias{constraint_bounds.matrix}
\alias{constraint_bounds.xts}
\alias{constraint_bounds.tbl_df}
\title{Bounds Constraint}
\usage{
constraint_bounds(.data, .min, .max)

\method{constraint_bounds}{default}(.data, .min, .max)

\method{constraint_bounds}{default}(.data, .min, .max)

\method{constraint_bounds}{matrix}(.data, .min, .max)

\method{constraint_bounds}{xts}(.data, .min, .max)

\method{constraint_bounds}{tbl_df}(.data, .min, .max)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{.tibble}.}

\item{.min}{A number with the minimum allowed weight.}

\item{.max}{A number with the maximum allowed weight.}
}
\value{
A list.
}
\description{
Bounds Constraint
}
\examples{
constraint_bounds(EuStockMarkets, 0, 1)
}
