% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_box}
\alias{constraint_box}
\alias{constraint_box.default}
\alias{constraint_box.matrix}
\alias{constraint_box.xts}
\alias{constraint_box.tbl_df}
\title{Budget Constraint}
\usage{
constraint_box(.data, .min, .max)

\method{constraint_box}{default}(.data, .min, .max)

\method{constraint_box}{matrix}(.data, .min, .max)

\method{constraint_box}{xts}(.data, .min, .max)

\method{constraint_box}{tbl_df}(.data, .min, .max)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{.tibble}.}

\item{.min}{A \code{number} with the minimum allowed allocation per asset.}

\item{.max}{A \code{number} with the maximum allowed allocation per asset.}
}
\value{
Not defined yet.
}
\description{
Helper that creates a box constraint for portfolio optimization.
}
\examples{
mtx <- matrix(EuStockMarkets, ncol = 4)
colnames(mtx) <- colnames(EuStockMarkets)

#constraint_box(mtx, .min = 0, .max = 1)
}
