% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_budget}
\alias{constraint_budget}
\alias{constraint_budget.default}
\alias{constraint_budget.matrix}
\alias{constraint_budget.xts}
\alias{constraint_budget.tbl_df}
\title{Budget Constraint}
\usage{
constraint_budget(.data, .investment = 1)

\method{constraint_budget}{default}(.data, .investment = 1)

\method{constraint_budget}{matrix}(.data, .investment = 1)

\method{constraint_budget}{xts}(.data, .investment = 1)

\method{constraint_budget}{tbl_df}(.data, .investment = 1)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{.tibble}.}

\item{.investment}{A \code{number} with the percentage of wealth that should
be allocated to risky assets.}
}
\value{
Not defined yet.
}
\description{
Helper that creates a budget constraint for portfolio optimization.
}
\examples{
constraint_budget(EuStockMarkets, 1)
}
