% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_cardinality}
\alias{constraint_cardinality}
\alias{constraint_cardinality.default}
\alias{constraint_cardinality.matrix}
\alias{constraint_cardinality.xts}
\alias{constraint_cardinality.tbl_df}
\title{Turnover Constraint}
\usage{
constraint_cardinality(.data, .card = 1)

\method{constraint_cardinality}{default}(.data, .card = 1)

\method{constraint_cardinality}{matrix}(.data, .card = 1)

\method{constraint_cardinality}{xts}(.data, .card = 1)

\method{constraint_cardinality}{tbl_df}(.data, .card = 1)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{.tibble}.}

\item{.card}{A \code{integer} with the minimum number of assets to be included
in the optimization process.}
}
\value{
Not defined yet.
}
\description{
Helper that creates a turnover constraint for portfolio optimization.
}
\examples{
constraint_cardinality(EuStockMarkets, .card = 2)
}
