% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_turnover}
\alias{constraint_turnover}
\alias{constraint_turnover.default}
\alias{constraint_turnover.matrix}
\alias{constraint_turnover.tbl_df}
\title{Turnover Constraint}
\usage{
constraint_turnover(.data, .bench = NULL, .turn = 100)

\method{constraint_turnover}{default}(.data, .bench = NULL, .turn = 100)

\method{constraint_turnover}{matrix}(.data, .bench = NULL, .turn = 100)

\method{constraint_turnover}{matrix}(.data, .bench = NULL, .turn = 100)

\method{constraint_turnover}{tbl_df}(.data, .bench = NULL, .turn = 100)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{.tibble}.}

\item{.bench}{A \code{numeric} vector with the weights of the benchmark
allocation.}

\item{.turn}{A \code{number} with the maximum allowed turnover.}
}
\value{
Not defined yet.
}
\description{
Helper that creates a turnover constraint for portfolio optimization.
}
\examples{
constraint_turnover(EuStockMarkets, .turn = 10)
}
