% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{construct_optimization}
\alias{construct_optimization}
\title{Build the Optimization Infrastructure}
\usage{
construct_optimization(
  .data,
  .objective,
  .constraints,
  .bounds,
  .maximum = TRUE
)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{tibble}.}

\item{.objective}{The objective function.}

\item{.constraints}{The constraints list.}

\item{.bounds}{Bounds used for optimization}

\item{.maximum}{Is it a maximization problem?}
}
\value{
A list.
}
\description{
Build the Optimization Infrastructure
}
\examples{
#
}
