% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_rebalance_infrastructure.R
\name{construct_rebalance_infrastructure}
\alias{construct_rebalance_infrastructure}
\title{Build a Rebalance Infrastrucure}
\usage{
construct_rebalance_infrastructure(
  .data,
  .by = c("day", "week", "month", "year")
)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{tibble}.}

\item{.by}{A \code{string} with the periodicity to compute the rebalance. One of: \code{day}, \code{week}, \code{month} or \code{year}.}
}
\value{
A \code{tibble} of the \code{snoop_rebalance} class. Contains 3 columns:
\code{.date}, \code{.analysis} and \code{.assessment}.
}
\description{
Prepares a \code{snoop_rolling} object before passing it through \code{\link{rebalance_portfolio}}.
}
\examples{
stocks <- tibble::tibble(
  time = as.Date('2009-01-01') + 0:99,
  X    = rnorm(100, 0, 1),
  Y    = rnorm(100, 0, 2),
  Z    = rnorm(100, 0, 4)
)

roll <- construct_rolling_infrastructure(stocks, .initial = 50)

const <- construct_rebalance_infrastructure(roll)
const

# Contains information under the hood:
class(const)
}
