% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_rolling_infrastructure.R
\name{construct_rolling_infrastructure}
\alias{construct_rolling_infrastructure}
\title{Build a Rolling Infraestructure}
\usage{
construct_rolling_infrastructure(
  .data,
  .initial = 252,
  .assess = 1,
  .cumulative = FALSE,
  ...
)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{tibble}.}

\item{.initial}{A \code{integer}. The number of samples used for analysis/modeling in the initial resample.}

\item{.assess}{A \code{integer}. The number of samples used for each assessment resample.}

\item{.cumulative}{A \code{boolean}. Should the analysis resample grow beyond the size specified by \code{.initial} at each resample?.}

\item{...}{Additional arguments to be passed to \code{\link[rsample]{rolling_origin}}.}
}
\value{
A \code{tibble} of the \code{snoop_rolling} class. Contains 3 columns:
\code{.date}, \code{.analysis} and \code{.assessment}.
}
\description{
Constructs a tidy-resampling object for time-series data.
}
\examples{
stocks <- tibble::tibble(
  time = as.Date('2009-01-01') + 0:99,
  X    = rnorm(100, 0, 1),
  Y    = rnorm(100, 0, 2),
  Z    = rnorm(100, 0, 4)
)

construct_rolling_infrastructure(stocks, .initial = 50)
}
