% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_performance.R
\name{extract_statistics}
\alias{extract_statistics}
\title{Portfolio Statistics}
\usage{
extract_statistics(.data, .cost = 0)
}
\arguments{
\item{.data}{A \code{tibble} from the \code{snoop_rolling}.}

\item{.cost}{A code{double} with the fixed cost per period of time.}
}
\value{
A \code{tibble} with 8 new columns: \code{.weights}, \code{.return} and \code{.volatility},
\code{.skewness} and \code{.kurtosis}, \code{.value_at_risk}, \code{.expected_shortfall} and
\code{.ena}.
}
\description{
Computes the portfolio statistics for each period being considered.
}
\examples{
stocks <- tibble::tibble(
  time = as.Date('2009-01-01') + 0:99,
  X    = rnorm(100, 0, 1),
  Y    = rnorm(100, 0, 2),
  Z    = rnorm(100, 0, 4)
)

roll <- construct_rolling_infrastructure(stocks, .initial = 50)

rebal <- construct_rebalance_infrastructure(roll)

# Mean Variance Strategy
mu_sigma <- function(.data) {
  list(mu = colMeans(.data), sigma = stats::cov(.data))
}

compute_cov <- function(.data) stats::cov(as.matrix(.data))

optimal <- rebalance_portfolio(rebal, compute_cov, .strategy = "risk_parity")
optimal

extract_statistics(optimal)
}
