% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{find_optimum}
\alias{find_optimum}
\title{Solve an Optimization Problem}
\usage{
find_optimum(.opt, .solver, .control, ...)
}
\arguments{
\item{.opt}{An object of the `OP` class.}

\item{.solver}{A \code{character} with an appliable solver.}

\item{.control}{A \code{list} for the control parameters}

\item{...}{Addicional arguments.}
}
\value{
A \code{list}.
}
\description{
Solve a problem.
}
\examples{
#
}
