% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{mean_variance}
\alias{mean_variance}
\title{Mean-Variance Objective Function}
\usage{
mean_variance(.data, .lambda)
}
\arguments{
\item{.data}{A \code{list} of length 2 with the location and dispersion parameters.}

\item{.lambda}{A \code{double} with the risk-aversion parameter.}
}
\value{
A \code{list} of the \code{objective} class.
}
\description{
Constructs a quadratic objective function.
}
\examples{
x <- diff(log(EuStockMarkets))

mean_variance(list(mu = colMeans(x), sigma = cov(x)), 1)
}
