% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{minimum_variance}
\alias{minimum_variance}
\title{Minimum Variance Objective Function}
\usage{
minimum_variance(.data)
}
\arguments{
\item{.data}{A variance-covariance matrix.}
}
\value{
A \code{list} of the \code{objective} class.
}
\description{
Constructs a quadratic objective function.
}
\examples{
x <- diff(log(EuStockMarkets))

minimum_variance(.data = cov(x))
}
