% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{rebalance_portfolio}
\alias{rebalance_portfolio}
\title{A Rebalance Structure}
\usage{
rebalance_portfolio(
  .data,
  .fn,
  ...,
  .strategy = c("risk_parity", "mean_variance")
)
}
\arguments{
\item{.data}{A tabular (non-tidy) \code{tibble}.}

\item{.fn}{A function to compute the optimization strategy}

\item{...}{Additional arguments to be passed to \code{.fn}.}

\item{.strategy}{A \code{character} with the optimization technique to be implemented.
Currently one of: \code{risk_parity} or \code{mean_variance}.}
}
\value{
A \code{tibble}.
}
\description{
A Rebalance Structure
}
\examples{
stocks <- tibble::tibble(
  time = as.Date('2009-01-01') + 0:99,
  X    = stats::rnorm(100, 0, 1),
  Y    = stats::rnorm(100, 0, 2),
  Z    = stats::rnorm(100, 0, 4)
)

roll <- construct_rolling_infrastructure(stocks, .initial = 50)

rebal <- construct_rebalance_infrastructure(roll)

# Mean Variance Strategy
mu_sigma <- function(.data) {
  list(mu = colMeans(.data), sigma = stats::cov(.data))
}

rebalance_portfolio(rebal, mu_sigma, .strategy = "mean_variance")

# Risk Parity Strategy
#compute_cov <- function(.data) stats::cov(as.matrix(.data))

#rebalance_portfolio(rebal, compute_cov, .strategy = "risk_parity")
}
