% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{risk_parity}
\alias{risk_parity}
\title{Risk-Parity Objective Function}
\usage{
risk_parity(.data)
}
\arguments{
\item{.data}{A variance-covariance matrix.}
}
\value{
A \code{list} of the \code{objective} class.
}
\description{
Constructs a risk-parity objective function.
}
\examples{
x <- diff(log(EuStockMarkets))

risk_parity(cov(x))
}
