% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{Skewness of a Time-series}
\usage{
skewness(.invariant, .method = c("moment", "fisher"), .na_rm = FALSE)
}
\arguments{
\item{.invariant}{An univariate or multivariate time series.}

\item{.method}{A \code{character} string. One of: \code{"excess"}, \code{"moment"} or \code{"fisher"}.}

\item{.na_rm}{A \code{logical} value. Should missing values be removed? The default is \code{FALSE}.}
}
\value{
A tidy \code{tibble} with 2 columns.
}
\description{
This function computes the kurtosis of a given data set. It was written for
internal computations.
}
\keyword{internal}
